requirements.txt
Django==4.2.2
psycopg
python-vlc==3.0.18122
requests==2.31.0
mysqlclient==2.1.0
bs4==0.0.1
numpy==1.23.2
pandas==1.5.2
djangorestframework==3.14.0